requirements.txt
# requirements.txt
# Pinned versions to resolve protobuf conflicts
protobuf==3.20.3  # Exact version groq needs (within streamlit's range)

# Core packages with compatible versions
streamlit==1.32.0
python-dotenv==1.0.0

# Supporting packages
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
pydantic==2.6.0